requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-ckeditor==6.5.1
django-environ==0.9.0
django-filter==22.1
django-jalali==6.0.0
django-jquery==3.1.0
django-js-asset==2.0.0
django-modeltranslation==0.18.5
django-parler==2.3
django-parler-rest==2.2
django-rosetta==0.9.8
djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.10.1
idna==3.4
inflection==0.5.1
jdatetime==4.1.0
jsonschema==4.16.0
Pillow==9.3.0
polib==1.1.1
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
